back/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
ultralytics==8.0.196
opencv-python==4.8.1.78
Pillow==10.1.0
pymysql==1.1.0
cryptography==41.0.7
sqlalchemy==2.0.23
python-dotenv==1.0.0
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
